CustomSplashScreen 2.0.0 - Rewrite
- Rewrite basically the entire mod for more stability and compatibility
- Now uses MidnightLib for configuration
- Added spinning loading indicator
- Support for random background images located in /config/customsplashscreen/backgrounds
- Boss Bar loading bar is now fixed and can be changed in color
- Splash screen can now be previewed!

All previous configs need to be rewritten, but all options are still available.

+++ src/main/java/eu/midnightdust/customsplashscreen/CustomSplashScreenClient.java
@@ -1,8 +1,6 @@
 package eu.midnightdust.customsplashscreen;
 
 import eu.midnightdust.customsplashscreen.config.CustomSplashScreenConfig;
-import me.shedaniel.autoconfig.AutoConfig;
-import me.shedaniel.autoconfig.serializer.JanksonConfigSerializer;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.loader.api.FabricLoader;
 
@@ -10,39 +8,35 @@ import java.io.*;
 import java.nio.file.*;
 
 public class CustomSplashScreenClient implements ClientModInitializer {
-
-    public static CustomSplashScreenConfig CS_CONFIG;
     public static File CONFIG_PATH = new File(FabricLoader.getInstance().getConfigDir() + "/customsplashscreen");
-    private static Path BackgroundTexture = Paths.get(CONFIG_PATH + "/background.png");
-    private static Path MojangTexture = Paths.get(CONFIG_PATH + "/mojangstudios.png");
-    private static Path MojankTexture = Paths.get(CONFIG_PATH + "/mojank.png");
-    private static Path ProgressBarTexture = Paths.get(CONFIG_PATH + "/progressbar.png");
-    private static Path ProgressBarBackgroundTexture = Paths.get(CONFIG_PATH + "/progressbar_background.png");
+    public static final Path BackgroundTexture = Paths.get(CONFIG_PATH + "/background.png");
+    public static final Path WideLogoTexture = Paths.get(CONFIG_PATH + "/wide_logo.png");
+    public static final Path SquareLogoTexture = Paths.get(CONFIG_PATH + "/square_logo.png");
+    public static final Path ProgressBarTexture = Paths.get(CONFIG_PATH + "/progressbar.png");
+    public static final Path ProgressBarBackgroundTexture = Paths.get(CONFIG_PATH + "/progressbar_background.png");
 
     @Override
     public void onInitializeClient() {
-        AutoConfig.register(CustomSplashScreenConfig.class, JanksonConfigSerializer::new);
-        CS_CONFIG = AutoConfig.getConfigHolder(CustomSplashScreenConfig.class).getConfig();
+        CustomSplashScreenConfig.init("customsplashscreen", CustomSplashScreenConfig.class);
 
         if (!CONFIG_PATH.exists()) { // Run when config directory is nonexistant //
             CONFIG_PATH.mkdir(); // Create our custom config directory //
-
-            // Open Input Streams for copying the default textures to the config directory //
-            InputStream background = Thread.currentThread().getContextClassLoader().getResourceAsStream("background.png");
-            InputStream mojangstudios = Thread.currentThread().getContextClassLoader().getResourceAsStream("mojangstudios.png");
-            InputStream mojank = Thread.currentThread().getContextClassLoader().getResourceAsStream("mojank.png");
-            InputStream progressbar = Thread.currentThread().getContextClassLoader().getResourceAsStream("progressbar.png");
-            InputStream progressbarBG = Thread.currentThread().getContextClassLoader().getResourceAsStream("progressbar_background.png");
-            try {
-                // Copy the default textures into the config directory //
-                Files.copy(background,BackgroundTexture,StandardCopyOption.REPLACE_EXISTING);
-                Files.copy(mojangstudios,MojangTexture,StandardCopyOption.REPLACE_EXISTING);
-                Files.copy(mojank,MojankTexture,StandardCopyOption.REPLACE_EXISTING);
-                Files.copy(progressbar,ProgressBarTexture,StandardCopyOption.REPLACE_EXISTING);
-                Files.copy(progressbarBG,ProgressBarBackgroundTexture,StandardCopyOption.REPLACE_EXISTING);
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
+        }
+        // Open Input Streams for copying the default textures to the config directory //
+        InputStream background = Thread.currentThread().getContextClassLoader().getResourceAsStream("background.png");
+        InputStream wide = Thread.currentThread().getContextClassLoader().getResourceAsStream("wide_logo.png");
+        InputStream square = Thread.currentThread().getContextClassLoader().getResourceAsStream("square_logo.png");
+        InputStream progressbar = Thread.currentThread().getContextClassLoader().getResourceAsStream("progressbar.png");
+        InputStream progressbarBG = Thread.currentThread().getContextClassLoader().getResourceAsStream("progressbar_background.png");
+        try {
+            // Copy the default textures into the config directory //
+            if (!BackgroundTexture.toFile().exists()) Files.copy(background,BackgroundTexture,StandardCopyOption.REPLACE_EXISTING);
+            if (!WideLogoTexture.toFile().exists()) Files.copy(wide,WideLogoTexture,StandardCopyOption.REPLACE_EXISTING);
+            if (!SquareLogoTexture.toFile().exists()) Files.copy(square,SquareLogoTexture,StandardCopyOption.REPLACE_EXISTING);
+            if (!ProgressBarTexture.toFile().exists()) Files.copy(progressbar,ProgressBarTexture,StandardCopyOption.REPLACE_EXISTING);
+            if (!ProgressBarBackgroundTexture.toFile().exists()) Files.copy(progressbarBG,ProgressBarBackgroundTexture,StandardCopyOption.REPLACE_EXISTING);
+        } catch (Exception e) {
+            e.printStackTrace();
         }
     }
 }
